+++ backend/resources/__init__.py
@@ -1,0 +1,12 @@
+"""Resources of backend."""
+# Add new Resources here:
+from . import sample
+from . import projects
+from . import institutions
+
+# and here:
+BLUEPRINTS = [
+    sample.BP,
+    projects.BP,
+    institutions.BP,
+]
