+++ backend/resources/projects.py
@@ -1,6 +1,6 @@
 """Project Resource."""
 from flask import Blueprint, request, jsonify
-from backend.util.db import get_db
+from backend.database.db import get_db
 
 BP = Blueprint('projects', __name__, url_prefix='/api/projects')
