+++ backend_config.ini
@@ -1,4 +1,4 @@
-[DB]
+[Database]
 HOST=localhost
 USER=backend
 PASSWORD=softwareprojekt2020
