+++ requirements.txt
@@ -1,5 +1,6 @@
 Flask==1.1.2
 mysql-connector==2.2.9
+click==7.1.2
 sqlalchemy==1.3.17
 prospector[with_mypy,with_pyroma,with_vulture]
 bandit
